Move new FileTypesByExtNode class into new FileTypesByExtension class
